fix: refactor to avoid importing from current index
