+++ packages/auth/src/providers/cognito/tokenProvider/tokenProvider.ts
@@ -1,0 +1,10 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import { CognitoUserPoolsTokenProvider } from './CognitoUserPoolsTokenProvider';
+
+export const cognitoUserPoolsTokenProvider =
+	new CognitoUserPoolsTokenProvider();
+
+export const tokenOrchestrator =
+	cognitoUserPoolsTokenProvider.tokenOrchestrator;
