+++ packages/notifications/src/inAppMessaging/providers/pinpoint/types/inputs.ts
@@ -1,11 +1,11 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+import { IdentifyUserOptions } from './options';
 import {
-	IdentifyUserOptions,
 	InAppMessageConflictHandler,
 	OnMessageInteractionEventHandler,
-} from '.';
+} from './types';
 import {
 	InAppMessage,
 	InAppMessageInteractionEvent,
